So, it turns out that in stream mode the 'binary' library I was using has some... problems. Extreme memory usage and poor time performance mean that it isn't usable, even if it is more convenient.
I have refactored to be a bit more primitive. The code needs some major
tidying up and de-duplicating but at least it now works without requiring,
for e.g., 4gb of memory to unzip a 10mb zip file.

I imagine that the buffer mode of 'binary' may not be good either, so I will
consider removing that later.